+++ lib/variable/test/concat_test.cpp
@@ -32,16 +32,16 @@ TEST_F(ConcatTest, dimension_mismatch) {
   // Size mismatch
   EXPECT_THROW_DISCARD(
       concat(std::vector{base, base.slice({Dim::Y, 0, 1})}, Dim::X),
-      except::NotFoundError);
+      except::DimensionError);
   // Label mismatch
   auto xz = base;
   xz.rename(Dim::Y, Dim::Z);
   EXPECT_THROW_DISCARD(concat(std::vector{xz, base}, Dim::X),
-                       except::NotFoundError);
+                       except::DimensionError);
   // Missing label in first arg can (right now) not lead to braodcast
   EXPECT_THROW_DISCARD(
       concat(std::vector{base.slice({Dim::Y, 0}), base}, Dim::Z),
-      except::NotFoundError);
+      except::DimensionError);
   // Missing label in second arg, but this broadcasts
   EXPECT_NO_THROW_DISCARD(
       concat(std::vector{base, base.slice({Dim::Y, 0})}, Dim::Z));