+++ src/scipp/__init__.py
@@ -27,7 +27,7 @@ from .core import Variable, DataArray, Dataset, DType, Unit
 # Import errors
 from .core import BinEdgeError, BinnedDataError, CoordError, \
                          DataArrayError, DatasetError, DimensionError, \
-                         DTypeError, NotFoundError, SizeError, SliceError, \
+                         DTypeError, SizeError, SliceError, \
                          UnitError, VariableError, VariancesError
 # Import submodules
 from . import units